refactor(ui): move enterprise connection create call to Verify Domain step
Previously Continue on Select Provider both staged the chosen provider and
called `createEnterpriseConnection`. That meant the backend record was
created before the user had actually verified ownership of the domain we
were going to wire SSO into, which is why we needed an idempotency guard
in the first place.

Move the create call into Verify Domain's OTP `onResolve`, immediately
after `setPrimaryEmailAddress`. By the time we hit that branch the user
has verified the email and we've promoted it to primary, so the
connection is only created against a confirmed-owned domain. The
existing `enterpriseConnection` short-circuit inside `createConnection`
keeps the operation idempotent for users who navigate back through the
wizard.

Select Provider becomes a pure selection step: track local state, call
`setProvider`, advance. Trims the `__internal_useUserEnterpriseConnections`,
`useReverification`, `useSession`, `useUser`, `useCardState`, and error
routing it no longer needs. Tests are pared down to selection +
navigation behavior; the previous create-related assertions move out of
scope for this step.

A follow-up could add a `VerifyDomainStep.test.tsx` covering the new
create-on-resolve path. Skipped for this commit.
